requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.8.6
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5